test: move CompareEnumsHelper to the main_test.go file
